requirements.txt
Django==3.1.7
djangorestframework==3.12.4
django-cors-headers==3.7.0
django-extensions==3.1.2
django-rest-swagger==2.2.0
django-jenkins==0.110.0
validation==0.6.1
pika==1.2.0
numpy==1.19.5
paho-mqtt==1.5.1